feat: Add support for generator options
This change introduces a new `options` field for custom generators,
allowing extra parameters to be passed to `protoc` plugins via the
`--<name>_opt` flag.

This is useful for generators that accept additional configuration. For
example, `protoc-gen-ts_proto` can be configured with options like
`esModuleInterop=true`.

The `options` value is a list of strings that is passed directly to the
protoc plugin. The README has been updated to document this new feature.