test: Add test for generator options
This commit introduces a test case for the recently added generator
`options` feature.

It verifies that the options are correctly formatted and passed as
`--<name>_opt` arguments to the `protoc` command. The test runs `hatch build`
with verbose output and asserts that the expected arguments are present
in the command's stderr.

The `build_wheel` test helper has been refactored to support running
builds in verbose mode.